Conditionally update alert repeat times
commit ea2f7ec1cf7c3b332f64c7e5a5667e80f58d31ee[1] removed setting of
'notification frequency' after changing what to evaluate. Prometheus
alerts relayed on that behavior and it has become impossible to create
them. alert_default_repeat_time sets the value to zero and the field.
is then hidden[2]. Since there is a model validation for the zero value
the alerts can't be created. A similar behavior for "hourly alerts"
changed - their default notification frequency should change to zero.

[1] https://github.com/ManageIQ/manageiq-ui-classic/pull/2601
[2] https://github.com/ManageIQ/manageiq-ui-classic/blob/983bd1d690cd684d8e0f730f2e170fc2e3d69c7f/app/views/miq_policy/_alert_details.html.haml#L131