toolbar_chooser - add ems_container to the list of toolbars using ?display=performance
without this, toolbar builder will use the `#{@display}_center` path, and try looking for the `performance_center` toolbar, which doesn't exist.

The change fixes the bug and uses a shared performance toolbar.

OTOH it is named *vm* performance, and the only button is visible only with realtime `typ`, which doesn't seem available in containers.

Fixes ManageIQ/manageiq-ui-classic#3145